+++ core/core-common/src/datatypes/push_notifications.rs
@@ -1,9 +1,10 @@
+use base64::{prelude::BASE64_STANDARD, Engine};
 use proton_api_core::services::proton::common::AuthId;
 use proton_crypto_account::{keys::PGPDeviceKey, proton_crypto::crypto::PGPProviderSync};
 use proton_crypto_notifications::{
     DecryptableNotification, NotificationError, PGPEncryptedNotification,
 };
-use secrecy::SecretString;
+use secrecy::{ExposeSecret, Secret, SecretString};
 use serde::{Deserialize, Serialize};
 use tracing::error;
 
@@ -11,12 +12,52 @@ pub use proton_crypto_notifications::DecryptedNotification;
 
 use crate::os::{KeyChainEntryKind, StoreInKeyChain};
 
-/// Device key stored in the keychain
-pub struct StoredDevicePrivateKey(SecretString);
+/// Device public key stored in the database
+pub struct StoredDevicePublicKey(String);
 
-impl AsRef<SecretString> for StoredDevicePrivateKey {
-    fn as_ref(&self) -> &SecretString {
+impl AsRef<String> for StoredDevicePublicKey {
+    fn as_ref(&self) -> &String {
         &self.0
+    }
+}
+impl From<String> for StoredDevicePublicKey {
+    fn from(value: String) -> Self {
+        Self(value)
+    }
+}
+impl From<StoredDevicePublicKey> for String {
+    fn from(value: StoredDevicePublicKey) -> Self {
+        value.0
+    }
+}
+
+/// Device key stored in the keychain
+pub struct StoredDevicePrivateKey(Secret<Vec<u8>>);
+
+impl AsRef<Secret<Vec<u8>>> for StoredDevicePrivateKey {
+    fn as_ref(&self) -> &Secret<Vec<u8>> {
+        &self.0
+    }
+}
+
+impl StoredDevicePrivateKey {
+    /// Takes raw bytes
+    ///
+    #[must_use]
+    pub fn with_bytes(value: Vec<u8>) -> Self {
+        Self(Secret::new(value))
+    }
+
+    #[must_use]
+    fn to_base64(&self) -> SecretString {
+        let key = self.0.expose_secret();
+        SecretString::new(BASE64_STANDARD.encode(key))
+    }
+
+    fn from_base64(val: &SecretString) -> Option<Self> {
+        let val = val.expose_secret();
+        let bytes = BASE64_STANDARD.decode(val).ok()?;
+        Some(Self::with_bytes(bytes))
     }
 }
 
@@ -26,11 +67,11 @@ impl StoreInKeyChain for StoredDevicePrivateKey {
     }
 
     fn from_stored_string(s: SecretString) -> Self {
-        Self(s)
+        Self::from_base64(&s).expect("Keychain contains invalid key")
     }
 
     fn to_stored_string(&self) -> SecretString {
-        self.0.clone()
+        self.to_base64()
     }
 }