+++ mail/mail-uniffi/CHANGELOG.md
@@ -20,9 +20,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-  - Key chain takes `OSKeyChainEntryKind` parameter. Key chain implementation must support
+  - [ET-2367] Key chain takes `OSKeyChainEntryKind` parameter. Key chain implementation must support
      more than one entry at the time.
      - get -> load method rename
+  - [ET-2368] Saving registered device is now generating new device token key pair if necessary.
+  - [ET-2368] Registering device is now using device key instead of user key.
 
 ## [0.54.4] - 2025-03-06